COMPRESS-93: Support for alternative ZIP compression methods
Override the ZipEntry compression method getter and setter in ZipArchiveEntry. This prevents the IllegalArgumentException that was thrown whenever a zip entry with an unsupported compression method was encountered. With this change an IOException is thrown instead only if or when the entry is actually being read or written. In addition, a public ZipArchiveEntry.isSupportedCompressionMethod() method is introduced so client code can choose to explicitly skip reading ZIP entries with unsupported compression methods.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@890088 13f79535-47bb-0310-9956-ffa450edef68