COMPRESS-93: Support for alternative ZIP compression methods
Fix the Maven221MultiVolumeTest failure by not using skip() to skip bytes. The FileInputStream implementation of skip() will happily skip past the end of the file, so we never see the EOFException that Maven221MultiVolumeTest expects to see when closing a truncated entry.

Also added an explicit test case for the ability to skip entries with unsupported compression methods.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@890110 13f79535-47bb-0310-9956-ffa450edef68