COMPRESS-290 Fixed error message with large groupid
This is a bit of a simple solution to the issue, since there are obviously lots of other options that
could have similar updates. The reality is that most recent macs that are initialized to run in corporate
networks tend to get large GID's for the users. So I just fixed the one we actually have complaints about

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1651417 13f79535-47bb-0310-9956-ffa450edef68