remove powermockito tests for Java5 compatibility
the 1.6.x line of powermock is compiled with -target 1.6 and cannot be
used by our Java5 based build systems. powermock 1.5 is incompatible
with the version of JUnit we use.

I intend to start a vote about moving compress to Java6 after the 1.11
release - at which point the test will get added back.