experimental API for generic compressor for LZ77-like alhorithms
Right now I'm experimenting with the API as a building block for
adding write support to Snappy. The hard part of LZ77 (finding
matches for back-references) is generic enough to extract it.

If this works out it may help us implementing compressors for Snappy
or LZ4 and with quite a bit of additional work for Brotli or
DEFLATE64 (which use additional encoders on top of LZ77).