adapt Zip64SupportIT to #10
when we added ZIP64 support we excluded the extra field from the
central directory entry if it turned out to be superfluous - even with
Zip64Model.Always.

The PR #10 introduced with Compress 1.11 changes this to make Always
mean "always", but we forgot to adapt this test. Unfortunately nobody
has run the ITs in more than fourteen months.

Many thanks to Bruno P. Kinoshita for finding this.