Site nits (mostly formatting)
the "API Docs" is sorted from most updated to last updated; in this sense, git is often more updated than the "Latest release" and, therefore, is conceptually the most up-to-date documentation.

Update tar.xml

Fix broken link

Code formatting nits

Format a few class references as code for better visual highlight.

Code block nit

Rework code block to a CDATA section, in order to avoid the need to escape XML entities (as was being done) and also for consistency with other code blocks.