COMPRESS-658: fix formatting the lowest expressable DOS time
Since COMPRESS-613, we add an 'extra time data' field when the ZIP entry
timestamp is outside the range the DOS timestamp used in other parts of the
header can handle.

Before, it incorrectly also added this header when the ZIP entry timestamp
was exactly the lowest possible timestamp expressable as a DOS timestamp.

This had the consequence that, when reading a ZIP file with such entries and
writing it back again, ZipEntry would 'invent' an 'extra time data' field
where none was needed.

With this change, it correctly considers the lowest expressable date as
expressable, and avoids inventing the extra time data field in this case.