Separate ArchiveStreamConstants from ArchiveStreamFactory
Many archivers rely on the constants defined in ArchiveStreamFactory,
and ArchiveStreamFactory relies on the individual archiver classes.
This is currently a build cycle. By extracting the constants to an
ArchiveStreamConstants interface, this build cycle is broken.

This work enables us to build the individual archiver libraries
independently, which reduces binary size and the potential attack
surface in the case of vulnerabilities.