feat: Introduce builders for all ArchiveInputStream implementations (#706)
* feat: add generic builder for `ArchiveInputStream`

Introduce a generic builder to configure options common to all `ArchiveInputStream` implementations, providing a unified way to introduce shared parameters.

* feat: add builder for `ArArchiveInputStream`

Introduce a builder for `ArArchiveInputStream` and refactor existing
constructor calls to use it.

* feat: add builder for `ArjArchiveInputStream`

Introduce a builder for `ArjArchiveInputStream` and refactor existing
constructor calls to use it.

* feat: add builder for `CpioArchiveInputStream`

Introduce a builder for `CpioArchiveInputStream` and refactor existing
constructor calls to use it.

* feat: add builder for `DumpArchiveInputStream`

Introduce a builder for `DumpArchiveInputStream` and refactor existing
constructor calls to use it.

* feat: add builder for `JarArchiveInputStream` and `ZipArchiveInputStream`

Introduce a builder for `JarArchiveInputStream` and `ZipArchiveInputStream` and refactor existing
constructor calls to use it.

* feat: use builder instead of TarArchiveInputStream constructors

Refactor code to replace direct `TarArchiveInputStream` constructor calls
with the builder, improving consistency and readability.

* docs: add changelog entry

* fix: failing tests

* fix: Javadoc failure on JDK 8

* fix: rename `ArchiveInputStream.Builder` to `AbstractBuilder`

* fix: Javadoc

* fix: rename `spliSplitSig`

* fix: rename `allowStoredEntriesWithDataDescriptor`

* fix: document `ZipArchiveInputStream` constructor

* Update action description

---------

Co-authored-by: Gary Gregory <garydgregory@users.noreply.github.com>